+++ src/main/java/de/uib/opsidatamodel/serverdata/CacheIdentifier.java
@@ -248,14 +248,9 @@ public enum CacheIdentifier {
 	DEPOT_TO_PRODUCT_TO_DEPENDENCY_INFOS("depot_to_product_to_dependency_infos"),
 
 	/**
-	 * return type {@code List<String>}
+	 * return type {@code Set<String>}
 	 */
 	SOFTWARE_LIST("software_list"),
-
-	/**
-	 * return type {@code NavigableMap<String, Integer>}
-	 */
-	SOFTWARE_TO_NUMBER("software_to_number"),
 
 	/**
 	 * return type {@code NavigableMap<String, SWAuditEntry>}