+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## [4.3.2.X] - 2024-XX-XXs
 ## Changed
+- Include software entry's ident to "Missing software entry" dialog's message to make apparent which software entry is missing
 - Use insecure connection in Messagebus, when certificate verification is disabled
 - Maintain consistent log levels for unspecified lines
 - Save user preferences on table's display fields
@@ -13,7 +14,6 @@
 - `MySQLdb.integrityError` when moving clients to NOT_ASSIGNED directory/group
 - Product filter disabled after a reload or switching client selection
 - "Failed actions" search and selection
-- The lower control panel in logviewer keeps disappearing on resize
 - The lower control panel in logviewer keeps disappearing on resize
 - For hardware devices with the same name in different hardware class, the information is shown incorrectly
 - Can't change table column visibility in product table after full reload